Move random_qty from .testing to .operator
